refactor: extract event sources into mpsc pipeline
